2021.14[CL] Cleaned up counting logic & added tests
The counting logic was, slightly, overcomplicated by having to
specifically add the last character of the initial input back in so it
was properly counted. I changed the compression function so that it
includes the pair `(last nil)` to simplify this logic. No more special
case, parameter removed from `minmax-hash`.