2022.02[C++] - Works, had some fun, not stellar
Had some fun with booleans becoming 0/1 implicitly for the one line
scoring routine. The parsing is a hack, constructing a complex cons
structure of `(rps . (rps . rps))`. I didn't care for that, but it
worked and let me run through it in a single pass for parsing and two
passes for scoring. Scoring could be reduced to a single pass tracking
both scores.